% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_set.r
\docType{data}
\name{canines}
\alias{canines}
\title{canines data set}
\format{
A data.frame with 8 columns
\describe{
\item{Chien}{Name of the dog}
\item{Taille}{Size of the dog}
\item{Poids}{Weight of the dog}
\item{Velocite}{Velocity of the dog}
\item{Intelligence}{Intelligence of the dog}
\item{Affection}{Affection of the dog}
\item{Agressivite}{Aggressiveness of the dog}
\item{Fonction}{class of individuals}
}
}
\source{
http://tutoriels-data-mining.blogspot.com/2008/03/afcm-races-canines.html
}
\usage{
canines
}
\description{
Data with qualitative variables, description of some dog
}
\keyword{datasets}
