% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/univ_multi_fct.r
\name{data_manager}
\alias{data_manager}
\title{Create an object "ClusteringData"}
\usage{
data_manager(df, groupe)
}
\arguments{
\item{df}{a data frame (quantitative variable with type 'numeric')}

\item{groupe}{a vector with the value of the group for each individual}
}
\value{
an object of "ClusteringData" class
}
\description{
Create an object "ClusteringData"
}
\examples{
\dontrun{
objet<-data_manager(apb,apb_class)}

}
