% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/davies_bouldin.R
\name{davies_bouldin}
\alias{davies_bouldin}
\title{davies_bouldin function}
\usage{
davies_bouldin(X, clust, distance = "euclidean")
}
\arguments{
\item{X}{data_frame or matrix of features must be numeric}

\item{clust}{vector of result of the cluster, numeric or string}

\item{distance}{str of type of distance see {\link[stats]{dist}} \url{https://www.rdocumentation.org/packages/stats/versions/3.6.2/topics/dist}}
}
\value{
davies_bouldin index
}
\description{
function which calculates the Davies Bouldin coefficient where we can decide the type of distance
}
\references{
\url{https://en.wikipedia.org/wiki/Davies-Bouldin_index}
}
\author{
Baptiste Lanoue
}
