% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dunn.R
\name{dunn}
\alias{dunn}
\title{Dunn function}
\usage{
dunn(X, clust, distance = "euclidean")
}
\arguments{
\item{X}{data_frame or matrix of features must be numeric}

\item{clust}{vector of result of the cluster, numeric or string}

\item{distance}{str of type of distance see {\link[stats]{dist}} \url{https://www.rdocumentation.org/packages/stats/versions/3.6.2/topics/dist}}
}
\value{
A list with different dunn value.
\itemize{
  \item dunn_mean - Dunn coefficient with distance inter cluster by centroid.
  \item dunn_min - Dunn coefficient with distance inter cluster by the min.
  \item dunn_max - Dunn coefficient with distance inter cluster by the max.
}
}
\description{
function which calculates the Dunn coefficient with different distance inter neigbour cluster: centroid, min and max
we can also decide the type of distance
}
\references{
\url{https://en.wikipedia.org/wiki/Dunn_index}
}
\author{
Baptiste Lanoue
}
