% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/univ_multi_fct.r
\name{get_graph_car_univ}
\alias{get_graph_car_univ}
\title{Graph on classified data}
\usage{
get_graph_car_univ(objet, choice_graph = c("quali", "quanti"))
}
\arguments{
\item{objet}{object of 'ClusteringData' class}

\item{choice_graph}{vector for choose the graph with qualitative or quantitatives data, by default there are both}
}
\value{
some ggplot2 graph
}
\description{
Graph on classified data
}
\examples{
\dontrun{
objet<-data_manager(apb,apb_class)
#If object have qualitative and quantitatives variables all graph are display :
get_graph_car_univ(objet)
#If object have qualitative variable :
get_graph_car_univ(objet,choice_graph=c('quali'))
}

}
