% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/univ_multi_fct.r
\name{pca_graph}
\alias{pca_graph}
\title{Graph of a PCA on classified data}
\usage{
pca_graph(
  objet,
  axes = c(1, 2),
  graph = c("screeplot", "ind", "var"),
  colorvector = c("#00AFBB", "#E7B800", "#DE42DE", "#425CDE", "#8BDE42", "#DE9F42",
    "#DE425C", "#229A9A", "#863100", "#002486")
)
}
\arguments{
\item{objet}{object of 'ClusteringData' class}

\item{axes}{vector of two elements, containing the value of the dimensions represented in the graphs, by default dimension 1 and 2 are displayed}

\item{graph}{allows to choose the graph(s) you want to display}

\item{colorvector}{vector colors used for graphics}
}
\value{
some ggplot2 graph
}
\description{
Graph of a PCA on classified data
}
