% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/silhouette.R
\name{silhouette}
\alias{silhouette}
\title{Silhouette function}
\usage{
silhouette(X, clust, distance = "euclidean")
}
\arguments{
\item{X}{data_frame or matrix of features must be numeric}

\item{clust}{vector of result of the cluster, numeric or string}

\item{distance}{str of type of distance see {\link[stats]{dist}} \url{https://www.rdocumentation.org/packages/stats/versions/3.6.2/topics/dist}}
}
\value{
A list with Silhouette coefficient.
\itemize{
  \item s_coeff - Silhouette Coefficient of the result
  \item vect_point - Silhouette Coefficient for each observations of the data
  \item vect_clus -  Silhouette Coefficient for each cluster
}
}
\description{
function which calculates the Silhouette coefficient where we can decide the type of distance
}
\references{
\url{https://fr.wikipedia.org/wiki/Silhouette_(clustering)}
}
\author{
Baptiste Lanoue
}
