% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_set.r
\docType{data}
\name{small_apb}
\alias{small_apb}
\title{small_apb data set}
\format{
A data.frame with 8 columns
\describe{
\item{X}{Index of individuals}
\item{ï..Departement}{Department of the schooling}
\item{Type_filieres}{Type of sector of the schooling}
\item{Filiere}{Sector of the schooling}
\item{Taux_academie_origine}{Rate of students coming from their home academy}
\item{Taux_admis_boursier}{Grant rate}
\item{Bac_sans_menion}{average number of Student with a Bac without mention}
\item{Bac_mention_AB}{average number of Student with a Bac mention good enough}
\item{Bac_mention_B}{average number of Student with a Bac mention good}
\item{Bac_mention_TB}{average number of Student with a Bac mention very good}
\item{Taux_bac_mention_TB}{Rate of students with the baccalaureate mention very well}
\item{Taux_remplissage}{Schooling filling rate}
\item{Taux_fille}{Daughter rate of schooling}
\item{Taux_admis_voeu1}{Rate of admissions having made wishes 1 for the schooling}
\item{groupes_cah}{class of individuals}
}
}
\source{
invented data
}
\usage{
small_apb
}
\description{
Data frame with mixed data, description of student
}
\keyword{datasets}
