% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calinski.R
\name{variance_index}
\alias{variance_index}
\title{Variance_Index}
\usage{
variance_index(X, clust)
}
\arguments{
\item{X}{data_frame of features must be numeric}

\item{clust}{vector of result of the cluster, numeric or string}
}
\value{
A list with different internal criteria value.
\itemize{
  \item calinski_harabasz - calinski_harabasz coefficient
  \item ball_hall - Ball_hall coefficient.
  \item Hartigan -  Hartigan coefficient.
  \item Xu -  Xu coefficient.
}
}
\description{
function which calculates variance internal criteria : Calinski_Harabasz, Ball_hall, Hartigan and Xu"
}
\references{
\url{https://slideplayer.com/slide/6982424/}
}
\author{
Baptiste Lanoue
}
